+++ babel.config.js
@@ -1,20 +1,18 @@
 // for ESM don't transpile modules
 
 const envConfig = {
-  ignoreBrowserslistConfig: true,
   modules: 'commonjs',
-  targets: { node: true },
-  bugfixes: true,
 };
 
 if (process.env.ESM) {
   envConfig.modules = false;
+  envConfig.targets = { node: true };
+  envConfig.bugfixes = true;
 }
 
 if (process.env.CDN) {
   envConfig.modules = 'umd';
   envConfig.targets = null;
-  envConfig.ignoreBrowserslistConfig = false;
 }
 
 module.exports = {
@@ -28,9 +26,9 @@ module.exports = {
     test: {
       plugins: [require.resolve('babel-plugin-macros')],
     },
+    development: {
+      compact: false,
+    },
   },
-  plugins: [
-    require.resolve('@babel/plugin-proposal-class-properties'),
-    require.resolve('@babel/plugin-syntax-dynamic-import'),
-  ],
+  plugins: [require.resolve('@babel/plugin-proposal-class-properties')],
 };