+++ examples/monaco-graphql-webpack/babel.config.js
@@ -4,22 +4,11 @@ module.exports = {
     [
       require.resolve('@babel/preset-env'),
       {
-        // corejs: { version: 3, proposals: true },
-        // useBuiltIns: 'usage',
         targets: { browsers: ['last 2 chrome versions'] },
         bugfixes: true,
       },
     ],
-    require.resolve('@babel/preset-typescript'),
+    [require.resolve('@babel/preset-typescript'), {}],
   ],
-  plugins: [
-    require.resolve('@babel/plugin-syntax-dynamic-import'),
-    require.resolve('@babel/plugin-proposal-class-properties'),
-    [
-      '@babel/plugin-transform-runtime',
-      {
-        regenerator: true,
-      },
-    ],
-  ],
+  plugins: [require.resolve('@babel/plugin-proposal-class-properties')],
 };