+++ packages/graphiql/src/api/hooks/useOperation.ts
@@ -1,0 +1,6 @@
+import { useSessionContext } from '../providers/GraphiQLSessionProvider';
+
+export default function useOperation() {
+  const { operation } = useSessionContext();
+  return operation;
+}