+++ packages/graphiql/src/api/hooks/useSchema.ts
@@ -1,0 +1,6 @@
+import { useSchemaContext } from '../providers/GraphiQLSchemaProvider';
+
+export default function useSchema() {
+  const { schema } = useSchemaContext();
+  return schema;
+}