+++ packages/graphiql/src/cdn.ts
@@ -4,19 +4,10 @@
  *  This source code is licensed under the MIT license found in the
  *  LICENSE file in the root directory of this source tree.
  */
-import 'regenerator-runtime/runtime';
-
-import './css/app.css';
-import './css/codemirror.css';
-import './css/foldgutter.css';
-import './css/info.css';
-import './css/jump.css';
-import './css/lint.css';
-import './css/loading.css';
-import './css/show-hint.css';
-
-import './css/doc-explorer.css';
-import './css/history.css';
-
-export default GraphiQL;
 import { GraphiQL } from './components/GraphiQL';
+export default GraphiQL;
+
+if (typeof window !== 'undefined') {
+  // @ts-ignore
+  window.GraphiQL = GraphiQL;
+}