+++ packages/graphql-language-service-server/src/MessageProcessor.ts
@@ -6,16 +6,16 @@
  *  LICENSE file in the root directory of this source tree.
  *
  */
-
-import { readFileSync } from 'fs';
+import { readFileSync, existsSync, mkdirSync, writeFileSync } from 'fs';
 import { URL } from 'url';
-
-import type {
+import * as path from 'path';
+import {
   CachedContent,
   GraphQLCache,
   Uri,
   DefinitionQueryResult,
   GraphQLConfig,
+  GraphQLProjectConfig,
 } from 'graphql-language-service';
 
 import {
@@ -25,7 +25,7 @@ import {
 
 import { Range, Position } from 'graphql-language-service-utils';
 
-import {
+import type {
   CompletionParams,
   FileEvent,
   VersionedTextDocumentIdentifier,
@@ -33,7 +33,7 @@ import {
   DidOpenTextDocumentParams,
 } from 'vscode-languageserver-protocol';
 
-import {
+import type {
   Diagnostic,
   CompletionItem,
   CompletionList,
@@ -50,12 +50,18 @@ import {
   TextDocumentPositionParams,
   DocumentSymbolParams,
   SymbolInformation,
+  WorkspaceSymbolParams,
 } from 'vscode-languageserver';
+
+import type { UnnormalizedTypeDefPointer } from '@graphql-tools/load';
 
 import { getGraphQLCache } from './GraphQLCache';
 import { parseDocument } from './parseDocument';
 
 import { Logger } from './Logger';
+import { printSchema } from 'graphql';
+import { tmpdir } from 'os';
+import { loadConfig, GraphQLExtensionDeclaration } from 'graphql-config';
 
 type CachedDocumentType = {
   version: number;
@@ -70,16 +76,22 @@ export class MessageProcessor {
   _isInitialized: boolean;
   _willShutdown: boolean;
   _logger: Logger;
-  _extensions?: Array<(config: GraphQLConfig) => GraphQLConfig>;
+  _extensions?: GraphQLExtensionDeclaration[];
   _fileExtensions?: Array<string>;
   _parser: typeof parseDocument;
+  _tmpDir: string;
+  _tmpUriBase: string;
+  _tmpDirBase: string;
+  _loadConfigOptions?: Parameters<typeof loadConfig>;
 
   constructor(
     logger: Logger,
-    extensions?: Array<(config: GraphQLConfig) => GraphQLConfig>,
+    extensions?: GraphQLExtensionDeclaration[],
     config?: GraphQLConfig,
     parser?: typeof parseDocument,
     fileExtensions?: string[],
+    tmpDir?: string,
+    loadConfigOptions?: Parameters<typeof loadConfig>,
   ) {
     this._textDocumentCache = new Map();
     this._isInitialized = false;
@@ -89,6 +101,15 @@ export class MessageProcessor {
     this._fileExtensions = fileExtensions;
     this._graphQLConfig = config;
     this._parser = parser ?? parseDocument;
+    this._tmpDir = tmpDir || tmpdir();
+    this._tmpDirBase = path.join(this._tmpDir, 'graphql-language-service');
+    this._tmpUriBase = path.join('file://', this._tmpDirBase);
+    if (loadConfigOptions) {
+      this._loadConfigOptions = loadConfigOptions;
+    }
+    if (!existsSync(this._tmpDirBase)) {
+      mkdirSync(this._tmpDirBase);
+    }
   }
 
   async handleInitializeRequest(
@@ -120,12 +141,13 @@ export class MessageProcessor {
         '`--configDir` option or `rootPath` argument is required.',
       );
     }
-
     this._graphQLCache = await getGraphQLCache(
       rootPath,
       this._parser,
       this._extensions,
       this._graphQLConfig,
+      this._fileExtensions,
+      // this._loadConfigOptions,
     );
     this._languageService = new GraphQLLanguageService(this._graphQLCache);
 
@@ -141,6 +163,8 @@ export class MessageProcessor {
         messageType: 'initialize',
       }),
     );
+    const config = this._graphQLCache.getGraphQLConfig();
+    await this._cacheAllProjectFiles(config);
 
     return serverCapabilities;
   }
@@ -161,6 +185,7 @@ export class MessageProcessor {
     const diagnostics: Diagnostic[] = [];
 
     let contents: CachedContent[] = [];
+
     // Create/modify the cached entry if text is provided.
     // Otherwise, try searching the cache to perform diagnostics.
     if ('text' in textDocument && textDocument.text) {
@@ -168,7 +193,7 @@ export class MessageProcessor {
       // Only run the below function if text is passed in.
       contents = this._parser(textDocument.text, uri, this._fileExtensions);
 
-      this._invalidateCache(textDocument, uri, contents);
+      await this._invalidateCache(textDocument, uri, contents);
     } else {
       const cachedDocument = this._getCachedDocument(textDocument.uri);
       if (cachedDocument) {
@@ -242,15 +267,15 @@ export class MessageProcessor {
       this._fileExtensions,
     );
     // If it's a .graphql file, proceed normally and invalidate the cache.
-    this._invalidateCache(textDocument, uri, contents);
+    await this._invalidateCache(textDocument, uri, contents);
 
     const cachedDocument = this._getCachedDocument(uri);
     if (!cachedDocument) {
       return null;
     }
 
-    this._updateFragmentDefinition(uri, contents);
-    this._updateObjectTypeDefinition(uri, contents);
+    await this._updateFragmentDefinition(uri, contents);
+    await this._updateObjectTypeDefinition(uri, contents);
 
     // Send the diagnostics onChange as well
     const diagnostics: Diagnostic[] = [];
@@ -457,8 +482,8 @@ export class MessageProcessor {
           const text: string = readFileSync(new URL(uri).pathname).toString();
           const contents = this._parser(text, uri, this._fileExtensions);
 
-          this._updateFragmentDefinition(uri, contents);
-          this._updateObjectTypeDefinition(uri, contents);
+          await this._updateFragmentDefinition(uri, contents);
+          await this._updateObjectTypeDefinition(uri, contents);
 
           const diagnostics = (
             await Promise.all(
@@ -519,7 +544,12 @@ export class MessageProcessor {
     }
     const textDocument = params.textDocument;
     const position = params.position;
-
+    const project = this._graphQLCache
+      .getGraphQLConfig()
+      .getProjectForFile(textDocument.uri);
+    if (project) {
+      await this._cacheSchemaFilesForProject(project);
+    }
     const cachedDocument = this._getCachedDocument(textDocument.uri);
     if (!cachedDocument) {
       throw new Error(`${textDocument.uri} is not available.`);
@@ -564,10 +594,6 @@ export class MessageProcessor {
           };
         })
       : [];
-
-    const project = this._graphQLCache
-      .getGraphQLConfig()
-      .getProjectForFile(textDocument.uri);
 
     this._logger.log(
       JSON.stringify({
@@ -623,6 +649,219 @@ export class MessageProcessor {
   //    );
   // }
 
+  async handleWorkspaceSymbolRequest(
+    params: WorkspaceSymbolParams,
+  ): Promise<Array<SymbolInformation>> {
+    // const config = await this._graphQLCache.getGraphQLConfig();
+    // await this._cacheAllProjectFiles(config);
+    const documents = Array.from(this._textDocumentCache);
+    let symbols: SymbolInformation[] = [];
+    await Promise.all(
+      documents.map(async ([uri]) => {
+        const cachedDocument = this._getCachedDocument(uri);
+        if (!cachedDocument) {
+          throw new Error('A cached document cannot be found.');
+        }
+        const docSymbols = await this._languageService.getDocumentSymbols(
+          cachedDocument.contents[0].query,
+          uri,
+        );
+        symbols.push(...docSymbols);
+      }),
+    );
+    if (params.query !== '') {
+      symbols = symbols.filter(
+        symbol => symbol?.name && symbol.name.includes(params.query),
+      );
+    }
+
+    return symbols;
+  }
+
+  async _cacheSchemaText(uri: string, text: string, version: number) {
+    try {
+      const contents = this._parser(text, uri, this._fileExtensions);
+      if (contents.length > 0) {
+        await this._updateObjectTypeDefinition(uri, contents);
+        await this._invalidateCache({ version, uri }, uri, contents);
+      }
+    } catch (err) {
+      this._logger.error(err);
+    }
+  }
+  async _cacheSchemaFile(
+    uri: UnnormalizedTypeDefPointer,
+    project: GraphQLProjectConfig,
+  ) {
+    uri = uri.toString();
+
+    const isFileUri = existsSync(uri);
+
+    let version = 1;
+    if (isFileUri) {
+      const schemaUri = 'file://' + path.join(project.dirpath, uri);
+      const schemaDocument = this._getCachedDocument(schemaUri);
+
+      if (schemaDocument) {
+        version = schemaDocument.version++;
+      }
+      const schemaText = readFileSync(uri, { encoding: 'utf-8' });
+      this._cacheSchemaText(schemaUri, schemaText, version);
+    }
+  }
+  _getTmpProjectPath(
+    project: GraphQLProjectConfig,
+    prependWithProtocol: boolean = true,
+  ) {
+    const baseDir = this._graphQLCache.getGraphQLConfig().dirpath;
+    const baseName = path.basename(baseDir);
+    const basePath = path.join(this._tmpDirBase, baseName);
+    if (!existsSync(basePath)) {
+      mkdirSync(basePath);
+      mkdirSync(path.join(basePath, 'projects'));
+    }
+    const projectTmpPath = path.resolve(
+      path.join(basePath, 'projects', project.name),
+    );
+    if (!existsSync(projectTmpPath)) {
+      mkdirSync(projectTmpPath);
+    }
+    return prependWithProtocol ? 'file://' + projectTmpPath : projectTmpPath;
+  }
+  async _cacheSchemaFilesForProject(project: GraphQLProjectConfig) {
+    const schema = project?.schema;
+    const config = project?.extensions?.vscode;
+    /**
+     * By default, let's only cache the full graphql config schema.
+     * This allows us to rely on graphql-config's schema building features
+     * And our own cache implementations
+     * This prefers schema instead of SDL first schema development, but will still
+     * work with lookup of the actual .graphql schema files if the option is enabled,
+     * however results may vary.
+     *
+     * The default temporary schema file seems preferrable until we have graphql source maps
+     */
+    const cacheSchemaFiles = config?.cacheSchemaFiles || false;
+    const cacheConfigSchema = config?.cacheConfigSchema || true;
+
+    if (cacheConfigSchema) {
+      await this._cacheConfigSchema(project);
+    }
+    if (cacheSchemaFiles && schema) {
+      if (Array.isArray(schema)) {
+        Promise.all(
+          schema.map(async (uri: UnnormalizedTypeDefPointer) => {
+            await this._cacheSchemaFile(uri, project);
+          }),
+        );
+      } else {
+        const uri = schema.toString();
+        await this._cacheSchemaFile(uri, project);
+      }
+    }
+  }
+  /**
+   * Cache the schema as represented by graphql-config, with extensions
+   * from GraphQLCache.getSchema()
+   * @param project {GraphQLProjectConfig}
+   */
+  async _cacheConfigSchema(project: GraphQLProjectConfig) {
+    try {
+      const schema = await this._graphQLCache.getSchema(project.name);
+      if (schema) {
+        const schemaText = printSchema(schema, {
+          commentDescriptions: true,
+        });
+        // file:// protocol path
+        const uri = path.join(
+          this._getTmpProjectPath(project),
+          'schema.graphql',
+        );
+        // no file:// protocol for fs.writeFileSync()
+        const fsPath = path.join(
+          this._getTmpProjectPath(project, false),
+          'schema.graphql',
+        );
+
+        const cachedSchemaDoc = this._getCachedDocument(uri);
+
+        if (!cachedSchemaDoc) {
+          writeFileSync(fsPath, schemaText, {
+            encoding: 'utf-8',
+          });
+          await this._cacheSchemaText(uri, schemaText, 1);
+        }
+        // do we have a change in the getSchema result? if so, update schema cache
+        if (
+          cachedSchemaDoc &&
+          schemaText !== cachedSchemaDoc.contents[0].query
+        ) {
+          writeFileSync(fsPath, schemaText, {
+            encoding: 'utf-8',
+          });
+          await this._cacheSchemaText(
+            uri,
+            schemaText,
+            cachedSchemaDoc.version++,
+          );
+        }
+      }
+    } catch (err) {
+      this._logger.error(err);
+    }
+  }
+  /**
+   * Pre-cache all documents for a project.
+   *
+   * TODO: Maybe make this optional, where only schema needs to be pre-cached.
+   *
+   * @param project {GraphQLProjectConfig}
+   */
+  async _cacheDocumentFilesforProject(project: GraphQLProjectConfig) {
+    try {
+      const documents = await project.getDocuments();
+      return Promise.all(
+        documents.map(async document => {
+          if (!document.location || !document.rawSDL) {
+            return;
+          }
+          // build full system URI path with protocol
+          const uri = 'file://' + path.join(project.dirpath, document.location);
+          // i would use the already existing graphql-config AST, but there are a few reasons we can't yet
+          const contents = this._parser(
+            document.rawSDL,
+            uri,
+            this._fileExtensions,
+          );
+
+          await this._updateObjectTypeDefinition(uri, contents);
+          await this._updateFragmentDefinition(uri, contents);
+          await this._invalidateCache({ version: 1, uri }, uri, contents);
+        }),
+      );
+    } catch (err) {
+      this._logger.error(
+        `invalid/unknown file in graphql config documents entry:\n '${project.documents}'`,
+      );
+      this._logger.error(err);
+    }
+  }
+  /**
+   * This should only be run on initialize() really.
+   * Cacheing all the document files upfront could be expensive.
+   * @param config {GraphQLConfig}
+   */
+  async _cacheAllProjectFiles(config: GraphQLConfig) {
+    if (config?.projects) {
+      return Promise.all(
+        Object.keys(config.projects).map(async projectName => {
+          const project = await config.getProject(projectName);
+          await this._cacheSchemaFilesForProject(project);
+          await this._cacheDocumentFilesforProject(project);
+        }),
+      );
+    }
+  }
   _isRelayCompatMode(query: string): boolean {
     return (
       query.indexOf('RelayCompat') !== -1 ||
@@ -666,31 +905,33 @@ export class MessageProcessor {
 
     return null;
   }
-  _invalidateCache(
+  async _invalidateCache(
     textDocument: VersionedTextDocumentIdentifier,
     uri: Uri,
     contents: CachedContent[],
-  ): void {
+  ): Promise<Map<string, CachedDocumentType> | null> {
     if (this._textDocumentCache.has(uri)) {
       const cachedDocument = this._textDocumentCache.get(uri);
       if (
         cachedDocument &&
-        textDocument.version &&
+        textDocument &&
+        textDocument?.version &&
         cachedDocument.version < textDocument.version
       ) {
         // Current server capabilities specify the full sync of the contents.
         // Therefore always overwrite the entire content.
-        this._textDocumentCache.set(uri, {
+        return this._textDocumentCache.set(uri, {
           version: textDocument.version,
           contents,
         });
       }
-    } else if (textDocument.version) {
-      this._textDocumentCache.set(uri, {
+    } else if (textDocument?.version) {
+      return this._textDocumentCache.set(uri, {
         version: textDocument.version,
         contents,
       });
     }
+    return null;
   }
 }